Ensure align_vectors inputs are correct
If origin_direction or target_direction are shape (1, 3), it is
possible to raise a DeprecationWaring when declaring v_c.
I addressed this by adding a check of the shape of these inputs.
End users could call np.squeeze to ensure the vectors have the
correct shape.

If origin_direction or target_direction have some other shape, the
rest of the code won't work anyway.