Adding a temp_dir parameter to the VinaApp
Using the VinaApp for docking creates temporary files. In the original implementation those where just stored in /tmp which usually is fine.
However, if one wants to use VinaApp for (large scale) docking on a cluster there might not be a /tmp folder or the path to it might be somewher else. 
Usually a distributed file system is available as scratch which is perfectly fine with handling a lot of short living files. Therfore, one might want to or even need to have controll over where the tempory files are created. 
I propose to simply add a parameter temp_dir that defaults to None which we will set the NamedTemporaryFiles parameter dir to.
If dir is set to None it will just create a file under /tmp otherwise the given path is used.
This can be tested by doing:

```python
x=NamedTemporaryFile()
y=NamedTemporaryFile(dir=None)
z=NamedTemporaryFile(dir=".")
print(x.name)
print(y.name)
print(z.name)
```
which will create 3 different termporary files and print there full path.